UHF-12050: Stash
